backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic[email]>=1.8.2
typing-extensions>=4.0.0
requests>=2.26.0
python-dotenv>=0.19.0
web3>=5.24.0
eth-account>=0.5.9
PyJWT>=2.3.0
cryptography>=3.4.8
python-jose>=3.3.0
passlib>=1.7.4
aiohttp>=3.8.1
fastapi-cors>=0.0.6
pymongo>=4.0.0
bcrypt>=3.2.0
uuid>=1.30
python-dateutil>=2.8.2
email-validator>=1.1.3
qrcode>=7.3.1
pillow>=9.0.0
ipfshttpclient>=0.8.0
